+++ packages/common/src/utils/index.ts
@@ -1,2 +1,3 @@
-// eslint-disable-next-line import/prefer-default-export
 export { default as cn } from './shadcn/index';
+
+export { default as FetchWrapper } from './api/index';
